requirements.txt
Django==1.8.5
djangorestframework==3.4.4
python-nmap==0.6.1
netifaces==0.10.4
psutil==4.3.0
py-cpuinfo==0.2.3
ansible==2.1.1.0
celery==3.1.23
threadpool==1.3.2
markupsafe==0.23
amqp==2.0.3
anyjson==0.3.3
django-websocket-redis==0.4.6
gevent==1.1.2
greenlet==0.4.10
packaging==16.8
appdirs==1.4.0
paramiko==2.1.1